irequestor package renamed to requests
