+++ server/response_test.go
@@ -2,17 +2,17 @@ package server
 
 import (
 	"bytes"
-	"github.com/mysterium/node/requestor"
+	"github.com/mysterium/node/requests"
 	"github.com/stretchr/testify/assert"
 	"io/ioutil"
 	"net/http"
 	"testing"
 )
 
-var testRequestApiUrl = "http://testUrl"
+var testRequestAPIURL = "http://testUrl"
 
 func TestHttpErrorIsReportedAsErrorReturnValue(t *testing.T) {
-	req, err := requestor.NewGetRequest(testRequestApiUrl, "path", nil)
+	req, err := requests.NewGetRequest(testRequestAPIURL, "path", nil)
 	assert.NoError(t, err)
 
 	response := &http.Response{
@@ -29,7 +29,7 @@ type testResponse struct {
 
 func TestHttpResponseBodyIsParsedCorrectly(t *testing.T) {
 
-	req, err := requestor.NewGetRequest(testRequestApiUrl, "path", nil)
+	req, err := requests.NewGetRequest(testRequestAPIURL, "path", nil)
 	assert.NoError(t, err)
 
 	response := &http.Response{
